backend/requirements.txt
fastapi==0.95.0
boto3==1.26.0
requests==2.28.2
uvicorn==0.21.1
python-multipart
boto3
python-dotenv
sqlalchemy
psycopg2-binary
databases
asyncpg
alembic